Backend/requirements.txt
django
djangorestframework
django-cors-headers
django-multiselectfield
djangorestframework-simplejwt